Create a MS2ID object

createMS2ID processes the data contained in a CompDb object and saves the result as SQL and bigMemory files in a directory. Later on, that directory can be load as MS2ID object in order to be used repeatedly as reference library with the annotate function.

createMS2ID(
  name = "MS2ID",
  path = ".",
  cmpdb,
  noiseThresh = 0.01,
  calcSplash = TRUE,
  calcMmi = TRUE,
  overwrite = FALSE
)

Arguments

name
character(1) name of the directory where the files will be saved.

path
character(1) with the directory's path.

cmpdb
CompDb(1) object (with MS/MS spectra).

noiseThresh
A numeric defining the threshold used in the noise filtering of the MS/MS spectra. e.g. noiseThresh = 0.01 removes peaks with an intensity less than 1% of the base peak.

calcSplash
boolean(1) indicating if SPLASH values (when missing) must be calculated using the getSplash function.

calcMmi
boolean(1) indicating if compound monoisotopic mass values (when missing) must be calculated using the check_chemform function.

overwrite
boolean(1) indicating if the function can overwrite results.

Value

a character(1) naming the resulting directory (+ path)

See also