In-house database for high-throughput MS/MS annotation

The `MS2ID` class encapsulates a MS2 spectra database along with its metadata. Its internal structure allows to annotate query spectra - with the annotate function - using big spectra databases at high speed and low RAM requirements (typically 100 query spectra/min against a 1M5 spectra library). MS2ID class uses a SQLite database with the metadata and bigmemory files to store the spectra (i.e. peaks matrices) and their mass-charge index.

MS2ID(ms2idFolder)

Details

`MS2ID` objects should be created using the constructor function `MS2ID` providing the path where the library files are (i.e. SQL file name or any bigmemory file names)

General functions

`MS2ID()`: Constructor of a MS2ID object.

`annotate()`: function to annotate query spectra against a MS2ID library

Author

Josep M. Badia